Added dst_name index to relationship table. It dramatically speeds up enumerating edges with enumerateEdgesWithName:destinationKey:collection:usingBlock:.
Without this index it would use the name index and then have to scan through the remaining results. This isn't better than a linear scan in cases where most of the rows in the database all have the same name.

Incremented YAP_DATABASE_RELATIONSHIP_CLASS_VERSION so that the index gets added automatically.