Updates CloudKit limit error handling to account for CKReferences
The change adds an initial enumeration of the dirty records to find build a map of all the parent references that have references that are also changing.
The existing enumeration was modified to ensure change sets are as large as possible given the configurable limit AND keep parent/child references in the same change set.